Update RV3129.cpp
I found a bug associated with switching from 12-hour mode to 24-hour mode. Specifically, when converting from 12 hour mode, an hour value of 12 is not detected because the 12-hour select bit (bit 6) is set to 1. As a result the hour is never actually 12, so converting from something like 12:30 fails, and you get 24:30. The solution is to clear the 12-hour select bit prior to calculating the hour.